Completed most of the structure of the project
Created a new component FaceIcons to put inside the rolecard, set it up with some style for it to look good.

Added 8 sample images.

Added a fakedata json file to represent the data displayed in the RoleCard

Filled information for the RoleCard component, mainly the texts, the props, the styles and the new component FaceIcons

Still missing most of the styles and some functionalities